// web/app/model/SAV.ts
export interface SAV {
  BoxNames: string[];
  Personal: string;
  EventFlagCount?: string;
  MaxMoveID: string;
  MaxSpeciesID: string;
  MaxItemID: string;
  MaxAbilityID: string;
  Blocks?: string;
  AllBlocks?: string;
  Items?: string;
  MysteryGift?: string;
  PokeFinder?: string;
  Festa?: string;
  Daycare?: string;
  Records?: string;
  Played?: string;
  MyStatus?: string;
  Overworld?: string;
  Situation?: string;
  Config?: string;
  GameTime?: string;
  Misc?: string;
  Zukan?: string;
  BoxLayout?: string;
  BattleTree?: string;
  ResortSave?: string;
  FieldMenu?: string;
  Fashion?: string;
  Fame?: string;
  BattleAgency?: string;
  Extension: string;
  PKMExtensions: string;
  BlankPKM: string;
  PKMType: string;
  BoxCount: number;
  MaxEV: string;
  Generation: number;
  Context: string;
  EventWorkCount?: string;
  MaxStringLengthOT: string;
  MaxStringLengthNickname: string;
  MaxBallID: string;
  MaxGameID: string;
  Version: string;
  ID32: string;
  TID16: string;
  SID16: string;
  Game: GAMES;
  Gender: string;
  GameSyncIDSize?: string;
  GameSyncID?: string;
  Region?: string;
  Country?: string;
  ConsoleRegion?: string;
  Language: string;
  OT: string;
  MultiplayerSpriteID?: string;
  Money: string;
  PlayedHours: string;
  PlayedMinutes: string;
  PlayedSeconds: string;
  SecondsToStart: string;
  SecondsToFame: string;
  RecordCount?: string;
  Inventory: string;
  CurrentBox: string;
  BoxesUnlocked: number;
  BoxFlags: string;
  PartyCount: string;
  DaycareSeedSize: string;
  ChecksumsValid: string;
  ChecksumInfo: string;
  TimeStampCurrent?: string;
  TimeStampPrevious?: string;
  State: string;
  Metadata: string;
  PlayTimeString: string;
  MinGameID: string;
  HasWondercards: string;
  GiftAlbum: string;
  SlotCount: string;
  MaxMoney: string;
  MaxCoins: string;
  TrainerIDDisplayFormat: string;
  TrainerTID7: string;
  TrainerSID7: string;
  DisplayTID: string;
  DisplaySID: string;
  HasParty: string;
  PartyData: [Pokemon, Pokemon, Pokemon, Pokemon, Pokemon, Pokemon];
  HasDaycare: string;
  HasTwoDaycares: string;
  SIZE_BOXSLOT: string;
  MaxIV: string;
  PokeDex: Pokemon[];
  HasPokeDex: string;
  SeenCount: string;
  CaughtCount: string;
  PercentSeen: string;
  PercentCaught: string;
  HasBox: string;
  BoxSlotCount: number;
  IsStorageFull: string;
  HasBoxWallpapers: string;
  HasNamableBoxes: string;
  NationalDex?: string;
  SecurityKey?: string;
  ClockInitial?: string;
  ClockElapsed?: string;
  JoyfulJumpInRow?: string;
  JoyfulJump5InRow?: string;
  JoyfulJumpGamesMaxPlayers?: string;
  JoyfulJumpScore?: string;
  JoyfulBerriesScore?: string;
  JoyfulBerriesInRow?: string;
  JoyfulBerries5InRow?: string;
  BP?: string;
  BPEarned?: string;
  Coin?: string;
  PokeBlocks?: string;
  Swarm?: string;
  DefaultSwarms?: string;
  SwarmIndex?: string;
  EBerryName?: string;
  IsEBerryEngima?: string;
  WonderOffset?: string;
  WonderNews?: string;
  WonderCard?: string;
  WonderCardExtra?: string;
  MysteryData?: string;
  RecordMixingGift?: string;
  WaldaBackgroundColor?: string;
  WaldaForegroundColor?: string;
  WaldaIconID?: string;
  WaldaPatternID?: string;
  WaldaUnlocked?: string;
  HasBattleVideo?: string;
  BattleVideo?: string;
  SaveRevision?: string;
  SaveRevisionString?: string;
  Japanese?: string;
  Korean?: string;
  PlayedFrames?: string;
  Badges?: string;
  DexPIDUnown?: string;
  DexPIDSpinda?: string;
  DexUnownForm?: string;
  PokedexSort?: string;
  PokedexMode?: string;
  PokedexNationalMagicRSE?: string;
  PokedexNationalMagicFRLG?: string;
  ColosseumRaw1?: string;
  ColosseumRaw2?: string;
  ColosseumCoupons?: string;
  ColosseumPokeCouponTitleGold?: string;
  ColosseumPokeCouponTitleSilver?: string;
  ColosseumPokeCouponTitleBronze?: string;
  ColosseumReceivedAgeto?: string;
  ColosseumCouponsTotal?: string;
  HasUsedRSBOX?: string;
  RSBoxDepositEggsUnlocked?: string;
  HasReceivedWishmkrJirachi?: string;
  Accessor?: string;
  BoxInfo?: string;
  PartyInfo?: string;
  PokedexSave?: string;
  AdventureStart?: string;
  LastSaved?: string;
  Status?: string;
  EventWork?: string;
  Storage?: string;
  GiftRecords?: string;
  Captured?: string;
}