// web/app/model/GAMES.ts
export enum GAMES {
  Any = 0,
  // The following values are IDs stored within PKM data, and can also identify individual games.
  S = 1,
  R = 2,
  E = 3,
  FR = 4,
  LG = 5,
  CXD = 15,
  D = 10,
  P = 11,
  Pt = 12,
  HG = 7,
  SS = 8,
  W = 20,
  B = 21,
  W2 = 22,
  B2 = 23,
  X = 24,
  Y = 25,
  AS = 26,
  OR = 27,
  SN = 30,
  MN = 31,
  US = 32,
  UM = 33,
  GO = 34,
  RD = 35,
  GN = 36,
  BU = 37,
  YW = 38,
  GD = 39,
  SI = 40,
  C = 41,
  GP = 42,
  GE = 43,
  SW = 44,
  SH = 45,

  // HOME = 46,
  PLA = 47,
  BD = 48,
  SP = 49,
  SL = 50,
  VL = 51,

  // The following values are not actually stored values in pk data,
  // These values are assigned within PKHeX as properties for various logic branching.
  RB,
  RBY,
  GS,
  GSC,
  RS,
  RSE,
  FRLG,
  RSBOX,
  COLO,
  XD,
  DP,
  DPPt,
  HGSS,
  BATREV,
  BW,
  B2W2,
  XY,
  ORASDEMO,
  ORAS,
  SM,
  USUM,
  GG,
  SWSH,
  BDSP,
  SV,
  Gen1,
  Gen2,
  Gen3,
  Gen4,
  Gen5,
  Gen6,
  Gen7,
  Gen7b,
  Gen8,
  Gen9,
  StadiumJ,
  Stadium,
  Stadium2,
}