// web/app/model/BUTTONS.ts
enum BUTTONS {
  B = 0,
  A = 1,
  Y = 2,
  X = 3,
  L = 4,
  R = 5,
  ZL = 6,
  ZR = 7,
  MINUS = 8,
  PLUS = 9,
  L_STICK = 10,
  R_STICK = 11,
  UP = 12,
  DOWN = 13,
  LEFT = 14,
  RIGHT = 15,
  HOME = 16,
  CAPTURE = 17,
  RSTICK_UP = 18,
  RSTICK_DOWN = 19,
  RSTICK_LEFT = 20,
  RSTICK_RIGHT = 21,
}
export default BUTTONS;